Start testing code to split CFG blocks
