+++ src/cfg/SplitCFG.h
@@ -2,31 +2,11 @@
 #define _SPLIT_CFG_H_
 
 #include "clang/Analysis/CFG.h"
+#include "SplitCFGBlock.h"
 #include <vector>
 #include <unordered_map>
 
 namespace systemc_clang {
-class SplitCFGBlock {
- private:
-  clang::CFGBlock *block_;
-  bool has_wait_;
-  std::vector<std::pair<unsigned int, unsigned int> > split_elements_;
-
- private:
-  bool isWait(const clang::CFGElement &element) const;
-
- public:
-  SplitCFGBlock();
-  SplitCFGBlock(const SplitCFGBlock &from);
-
-  clang::CFGBlock *getCFGBlock() const;
-  std::size_t getSplitBlockSize() const;
-
-  void split_block(clang::CFGBlock *block);
-
-  void dump() const;
-};
-
 /// ===========================================
 /// SplitCFG
 /// ===========================================
